imbue_core/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imbue-core"
version = "0.0.9"
description = "Utilities for imbue-desktop."
readme = "README.md"
authors = [
  { name="Imbue", email="imbue@imbue.com" },
]
license = "MIT"
# NOTE: This list is replicated in sculptor/pyproject.toml, and the copy there must be kept in sync.
dependencies = [
    "anyio",
    "attrs",
    "boto3>=1.38.27",
    "cachetools",
    "cattrs",
    "diskcache>=5.6.3",
    "grpclib>=0.4.7",
    "httpx",
    "inline-snapshot",
    "loguru",
    "pathspec",
    "posthog==5.4.0",
    "prometheus-client>=0.20.0",
    "pydantic-settings",
    "pydantic>=2.11.4",
    "pygit2>=1.18.0",
    "pylint==3.2.6",
    "pygments>=2.0.0",
    "pyhumps",
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "syrupy",
    "python-gitlab>=4.5.0",
    "sentry-sdk",
    "tblib==2.0.0", # pinning because Hammers code relies on "get_locals"
    "tenacity>=8.2.2",
    "toml",
    "traceback-with-variables>=2.2.0",
    "typeid-python",
    "yasoo",

    "anthropic~=0.54", # forcing a newer version that is not reliant on old httpx
    "openai>=1.79.0",
    "tiktoken",
    "together",
    "groq>=0.18.0",
    "google-genai>=1.26.0",
]
requires-python = ">=3.11,<3.13"

[project.urls]
"homepage" = "https://imbue.com/"

[tool.setuptools.package-data]
"imbue" = ["py.typed"]

[tool.setuptools.packages.find]
include = ["imbue_core*"]

[dependency-groups]
dev = [
    "moto>=4.1.12",
]
